refactor: update NewVideoContentHeader tests to ensure buttons render correctly and add DownloadDialog and ShareDialog functionality
